% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python.R
\name{check_python}
\alias{check_python}
\title{Checks of Python requirements}
\usage{
check_python(env_name = "rhsbm_env")
}
\description{
If Python requirements are not met, user must install hSBM requirements manually.
}
