% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_stopwords.R
\name{customize_stopwords}
\alias{customize_stopwords}
\title{Extract empirical stopwords}
\usage{
customize_stopwords(corpus, stopwords_sample, stopwords_threshold)
}
\description{
Extract empirical stopwords
}
