% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{process_text}
\alias{process_text}
\title{Preprocessing of raw or lemmatized text data}
\usage{
process_text(
  src_docs,
  src_texts,
  target_docs = "titles_processed.txt",
  target_texts = "corpus_processed.txt",
  replace_from = NULL,
  replace_to = NULL,
  collocations = TRUE,
  min_nchar = NULL,
  remove_punct = TRUE,
  remove_symbols = TRUE,
  remove_numbers = TRUE,
  remove_url = TRUE,
  remove_separators = TRUE,
  split_hyphens = FALSE,
  padding = FALSE,
  stopwords = NULL,
  min_termfreq = NULL,
  max_termfreq = NULL,
  termfreq_type = c("count", "prop", "rank", "quantile"),
  min_docfreq = NULL,
  max_docfreq = NULL,
  docfreq_type = c("count", "prop", "rank", "quantile")
)
}
\description{
Preprocessing of raw or lemmatized text data
}
